backend/ocr_service/requirements.txt
easyocr==1.7.2
flask==3.1.1
pillow==11.2.1
